requirements.txt
ccxt
pandas
numpy
python-telegram-bot>=20.0
python-dotenv
ta
cryptography
pyyaml
